Handle protobuf maps (#824)
Protobuf maps are just repeated protobuf messages with only 2 fields
"key" and "value". Extract the map's key and use it in the series name
and skip adding series for "key" fields to reduce the number of non
useful series. Additionally don't include "value" in the series name for
the value of a map.